+++ lkj/src/support_func.h
@@ -4,7 +4,6 @@
 #include "GyverEncoder.h"
 #include "Keypad.h"
 #include <math.h>
-// #include <avr/pgmspace.h>
 
 //Пины для дислпея 
 #define TFT_DC   8     //datacomand 
@@ -39,18 +38,13 @@ struct st_freq_params{
   uint32_t int_part_of_frac_div;
 };
 
-
-
-
-
-
-int find_chdiv(uint64_t);                                            //определение значения делителя
+uint8_t find_chdiv(uint64_t);                                            //определение значения делителя
 uint16_t replace_bits_8_to_13(uint16_t, uint8_t);
 uint32_t get_fractional_part(float);                                         //отделяет цифры после запятой
 uint16_t low_16bit(uint32_t);                                             //выделяет младшие 16 бит из 32 битной переменной
 uint16_t high_16bit(uint32_t);                                            //выделяет старшие 16 бит из 32 битной переменной 
-void print_freq(uint64_t, int ,int);
-int get_number_of_characters(uint64_t);                                              //возвращает количество цифр в числе 
+void print_freq(uint64_t, uint8_t, uint8_t);
+uint8_t get_number_of_characters(uint64_t);                                              //возвращает количество цифр в числе 
 uint32_t calculation_of_pll_n(uint64_t);
 uint32_t calculation_of_pll_num(uint64_t);     
 void power_print(uint32_t);
